Change reference counter from 64-bits to 32-bits, helps br pred somehow.

+++ VTIL-Common/util/copy_on_write.hpp
@@ -91,7 +91,7 @@ namespace vtil
 	{
 		// Declare the object entry.
 		//
-		using object_entry =   std::pair<T, std::atomic<size_t>>;
+		using object_entry =   std::pair<T, std::atomic<uint32_t>>;
 
 		// Declare object allocator.
 		//
